requirements.txt
milvus==2.0.0
